scripts/requirements.txt
keyboard==0.13.5
mss==10.0.0
opencv_python==4.11.0.86
pandas==2.0.1
Pillow==9.5.0
Pillow==11.3.0
PyDirectInput==1.0.4
pytesseract==0.3.13
pywin32==306
